Update Map-view
Fixed Small typo and added link to JDK. The link that's given by the CLI is going to https://www.java.com/en/download/mac_download.jsp , but this one doesn't work. It keeps saying you didn't install it so it can't be the right one. The link I gave works.